requirements.txt
absl-py==2.1.0
ale-py==0.8.1
astunparse==1.6.3
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
baselines==0.1.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
Farama-Notifications==0.0.4
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gast==0.5.4
glfw==2.7.0
google-pasta==0.2.0
grpcio==1.62.1
gym==0.26.2
gym-notices==0.0.8
gym-zelda-1==0.2.2
gymnasium==0.29.1
h5py==3.11.0
idna==3.7
imageio==2.34.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
keras==3.2.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpi4py==3.1.5
mpmath==1.3.0
mujoco==2.2.0
namex==0.0.7
nes-py==8.2.1
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
progressbar2==4.4.2
protobuf==4.25.3
psutil==5.9.8
pygame==2.1.0
pyglet==1.5.21
pygments==2.17.2
PyOpenGL==3.1.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
rich==13.7.1
scipy==1.13.0
Shimmy==1.3.0
six==1.16.0
stable-baselines3==2.3.0
stable-retro==0.9.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
torch==2.2.2
tqdm==4.66.2
triton==2.2.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
zmq==0.0.0
